fix changing to np.uint8

+++ q01_10/001.py
@@ -9,7 +9,7 @@ import lib0110
 
 if __name__ == '__main__':
     img = cv2.imread("Gasyori100knock/Question_01_10/imori.jpg")
-    ans = lib0110.BGR2RGB(img)
+    ans = lib0110.BGR2RGB(img).astype(np.uint8)
 
     cv2.imshow("imori", ans)
     cv2.waitKey(0)
