+++ q01_10/009.py
@@ -4,7 +4,7 @@ import lib0110
 
 if __name__ == '__main__':
     img = cv2.imread("Gasyori100knock/Question_01_10/imori_noise.jpg")
-    ans = lib0110.Gaussian_filter(img)
+    ans = lib0110.Gaussian_filter(img).astype(np.uint8)
 
     cv2.imshow("imori", ans)
     cv2.waitKey(0)
