+++ q21_30/026.py
@@ -6,7 +6,7 @@ import matplotlib.pyplot as plt
 if __name__ == '__main__':
     img = cv2.imread("Gasyori100knock/Question_21_30/imori.jpg")
     
-    ans = lib2130.Bi_linear_interpolation(img, ax=1.5, ay=1.5)
+    ans = lib2130.Bi_linear_interpolation(img, ax=1.5, ay=1.5).astype(np.uint8)
     cv2.imshow("imori", ans)
     cv2.waitKey(0)
     cv2.imwrite("q21_30/026.jpg", ans)
